+++ R/set_remotes.R
@@ -11,11 +11,13 @@
 #' @examples
 #' # Find from vector of packages
 #' find_remotes(pkg = c("attachment", "desc", "glue"))
+#'
 #' # Find from Description file
 #' dummypackage <- system.file("dummypackage", package = "attachment")
 #' att_from_description(
 #' path = file.path(dummypackage, "DESCRIPTION")) %>%
 #' find_remotes()
+#'
 #' \dontrun{
 #' # For the current package directory
 #' att_from_description() %>% find_remotes()