+++ NEWS.md
@@ -27,7 +27,7 @@ Recommendation: Run `att_amend_desc(pkg_ignore = "x", update.config = TRUE)` if 
 ## Minor changes
 
 - a new parameters `check_if_suggests_is_installed` in `att_amend_desc()` allow not to check if suggested package is installed. (thanks to @yogat3ch)
-- `create_renv_for_prod()` dont look anymore for suggested packages
+- `create_renv_for_prod()` don't look anymore for suggested packages
 - Clean a maximum of temp directories after examples and unit tests
 
 # attachment 0.3.0